refactor: move feature configuration to commonskmm
