+++ app/src/main/java/org/dhis2/usescases/searchTrackEntity/SearchTEActivity.kt
@@ -38,7 +38,7 @@ import org.dhis2.bindings.isKeyboardOpened
 import org.dhis2.bindings.overrideHeight
 import org.dhis2.commons.Constants
 import org.dhis2.commons.extensions.closeKeyboard
-import org.dhis2.commons.featureconfig.data.FeatureConfigRepository
+import org.dhis2.mobile.commons.featureconfig.data.FeatureConfigRepository
 import org.dhis2.commons.filters.FilterItem
 import org.dhis2.commons.filters.FilterManager
 import org.dhis2.commons.filters.FilterManager.PeriodRequest
