+++ app/src/main/java/org/dhis2/usescases/teiDashboard/DashboardRepositoryImpl.kt
@@ -8,7 +8,7 @@ import io.reactivex.Single
 import io.reactivex.functions.Function
 import org.dhis2.bindings.profilePicturePath
 import org.dhis2.commons.data.ProgramConfigurationRepository
-import org.dhis2.commons.featureconfig.data.FeatureConfigRepository
+import org.dhis2.mobile.commons.featureconfig.data.FeatureConfigRepository
 import org.dhis2.commons.prefs.Preference
 import org.dhis2.commons.prefs.PreferenceProvider
 import org.dhis2.commons.resources.MetadataIconProvider
