+++ commons/build.gradle.kts
@@ -66,7 +66,7 @@ kapt {
 
 dependencies {
     implementation(fileTree(mapOf("dir" to "libs", "include" to listOf("*.jar"))))
-    implementation(project(":commonskmm"))
+    api(project(":commonskmm"))
 
     api(libs.dhis2.android.sdk) {
         this.isChanging = true
