+++ commons/src/main/java/org/dhis2/commons/featureconfig/di/FeatureConfigActivityModule.kt
@@ -3,7 +3,7 @@ package org.dhis2.commons.featureconfig.di
 import dagger.Module
 import dagger.Provides
 import org.dhis2.commons.di.dagger.PerActivity
-import org.dhis2.commons.featureconfig.data.FeatureConfigRepository
+import org.dhis2.mobile.commons.featureconfig.data.FeatureConfigRepository
 import org.dhis2.commons.featureconfig.ui.FeatureConfigViewModelFactory
 
 @Module
