+++ commons/src/main/java/org/dhis2/commons/featureconfig/di/FeatureConfigModule.kt
@@ -2,8 +2,8 @@ package org.dhis2.commons.featureconfig.di
 
 import dagger.Module
 import dagger.Provides
-import org.dhis2.commons.featureconfig.data.FeatureConfigRepository
-import org.dhis2.commons.featureconfig.data.FeatureConfigRepositoryImpl
+import org.dhis2.mobile.commons.featureconfig.data.FeatureConfigRepository
+import org.dhis2.mobile.commons.featureconfig.data.FeatureConfigRepositoryImpl
 import org.hisp.dhis.android.core.D2
 import org.hisp.dhis.android.core.D2Manager
