+++ commons/src/main/java/org/dhis2/commons/featureconfig/ui/FeatureConfigViewModel.kt
@@ -3,8 +3,8 @@ package org.dhis2.commons.featureconfig.ui
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import org.dhis2.commons.featureconfig.data.FeatureConfigRepository
-import org.dhis2.commons.featureconfig.model.FeatureState
+import org.dhis2.mobile.commons.featureconfig.data.FeatureConfigRepository
+import org.dhis2.mobile.commons.featureconfig.model.FeatureState
 
 class FeatureConfigViewModel(
     private val repository: FeatureConfigRepository,
