+++ commons/src/main/java/org/dhis2/commons/service/SessionManagerModule.kt
@@ -2,7 +2,7 @@ package org.dhis2.commons.service
 
 import dagger.Module
 import dagger.Provides
-import org.dhis2.commons.featureconfig.data.FeatureConfigRepository
+import org.dhis2.mobile.commons.featureconfig.data.FeatureConfigRepository
 import org.dhis2.commons.prefs.PreferenceProvider
 import org.dhis2.commons.schedulers.SchedulerProvider
 import org.hisp.dhis.android.core.D2
