+++ commonskmm/src/commonMain/kotlin/org/dhis2/mobile/commons/featureconfig/data/FeatureConfigRepository.kt
@@ -1,7 +1,7 @@
-package org.dhis2.commons.featureconfig.data
+package org.dhis2.mobile.commons.featureconfig.data
 
-import org.dhis2.commons.featureconfig.model.Feature
-import org.dhis2.commons.featureconfig.model.FeatureState
+import org.dhis2.mobile.commons.featureconfig.model.Feature
+import org.dhis2.mobile.commons.featureconfig.model.FeatureState
 
 interface FeatureConfigRepository {
     val featuresList: List<FeatureState>
