+++ commonskmm/src/commonMain/kotlin/org/dhis2/mobile/commons/featureconfig/model/FeatureState.kt
@@ -1,4 +1,4 @@
-package org.dhis2.commons.featureconfig.model
+package org.dhis2.mobile.commons.featureconfig.model
 
 data class FeatureState(
     val feature: Feature,
