+++ commonskmm/src/desktopMain/kotlin/org/dhis2/mobile/commons/di/CommonsModule.desktop.kt
@@ -1,6 +1,9 @@
 package org.dhis2.mobile.commons.di
 
 import org.dhis2.mobile.commons.data.ValueParser
+import org.dhis2.mobile.commons.featureconfig.data.FeatureConfigRepository
+import org.dhis2.mobile.commons.featureconfig.model.Feature
+import org.dhis2.mobile.commons.featureconfig.model.FeatureState
 import org.dhis2.mobile.commons.model.internal.ValueInfo
 import org.koin.core.module.Module
 import org.koin.dsl.module
@@ -8,6 +11,17 @@ import org.koin.dsl.module
 actual val commonsModule: Module
     get() =
         module {
+            single<FeatureConfigRepository> {
+                object : FeatureConfigRepository {
+                    override val featuresList: List<FeatureState>
+                        get() = emptyList()
+
+                    override fun updateItem(featureState: FeatureState) = Unit
+
+                    override fun isFeatureEnable(feature: Feature) = false
+                }
+            }
+
             single {
                 object : ValueParser {
                     override suspend fun getValueInfo(
