+++ dhis_android_analytics/src/main/java/dhis2/org/analytics/charts/Charts.kt
@@ -3,7 +3,7 @@ package dhis2.org.analytics.charts
 import android.content.Context
 import dhis2.org.analytics.charts.data.Graph
 import dhis2.org.analytics.charts.ui.OrgUnitFilterType
-import org.dhis2.commons.featureconfig.data.FeatureConfigRepository
+import org.dhis2.mobile.commons.featureconfig.data.FeatureConfigRepository
 import org.dhis2.commons.resources.ColorUtils
 import org.dhis2.commons.viewmodel.DispatcherProvider
 import org.hisp.dhis.android.core.D2
